+++ CHANGELOG.md
@@ -6,11 +6,15 @@ All notable changes to this project will be documented in this file, following t
 ### Added
 - MMTF namespace
 - MMTF.decode convenience function that includes msgpack deserialization
+- MMTF.traverse function to loop over the structure hierarchy
 
 ### Changed
 - decoupled serialization from decoding, no more msgpack deserialization in `decodeMmtf`
 - renamed mmtf-decode-helpers to mmtf-utils
 - switched to UMD format for distribution files
+
+### Removed
+- MmtfIterator, use traversal function instead
 
 
 ## v0.1.0 - 2016-04-13